feat: interpret functions
Because of the comma operator it is not possible to parse expression
arguments in the function arguments list. comma operator has higher
presidence than expression, so parsing will return only 1 argument
(which is a comma expression) regardless of how many arguments are
passed in. This results in an error when checked for arity and argument
count equality. Therefore, when parsing function arguments the
interpreter must look for expressions with higher presidence than the
comma operator.